requirements.txt
Django
mysqlclient
python-dotenv
Pillow
requests
gunicorn
whitenoise
